[Feature:Autograding] Config option to keep `.git`

+++ autograder/submitty_autograding_shipper.py
@@ -1010,9 +1010,25 @@ def checkout_vcs_repo(config, my_file):
                 "credentials.\n",
                 file=f)
 
-    # remove the .git directory (storing full history and metafiles)
-    git_path = os.path.join(checkout_path, ".git")
-    shutil.rmtree(git_path, ignore_errors=True)
+    form_json_file = os.path.join(course_dir, 'config', 'form', f'form_{obj["gradeable"]}.json')
+    with open(form_json_file, 'r') as fj:
+        form_json = json.load(fj)
+    gradeable_json_file = os.path.join(form_json["config_path"], "config.json")
+    with open(gradeable_json_file, 'r') as gj:
+        gradeable_json = json.load(gj)
+
+    try:
+        # We explicitly compare equality with True because otherwise,
+        # Python might implicitly coerce a “truthy” value
+        # (such as any non-empty string, including the string "False")
+        # to True in an if statement
+        keep_git_directory = gradeable_json["keep_git_directory"] is True
+    except KeyError:
+        keep_git_directory = False
+    if not keep_git_directory:
+        # remove the .git directory (storing full history and metafiles)
+        git_path = os.path.join(checkout_path, ".git")
+        shutil.rmtree(git_path, ignore_errors=True)
 
     # calculate total file size, and remove symlinks
     (checkout_size, checkout_included_symlinks) = calculate_size_cleanup_symlinks(checkout_path)
